feat: route `/` and `/:categories/posts`
By clicking on each category, users can view list posts of particular category (this list will be implemented next commit)

In `index.js`, add a pathless `Route` for `App` component to add `location` object to props of the `App`.
```
<Route component={App}/>
```
*Note: learn more about [Update Blocking]("https://github.com/ReactTraining/react-router/blob/master/packages/react-router/docs/guides/blocked-updates.md")

In `Category.js`, add a path to `Link`
```
<Link to={`/${props.categoryPath}/posts`}>{props.categoryName}</Link>
```

In `App.js`, set param to `Route`
```
<Route exact path='/:categories/posts' component={PostsByCategory}/>
```

`PostsByCategory` is a new component which will be implemented next commit.